server/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
httpx==0.28.1
pyyaml==6.0.2
python-multipart==0.0.20

# File Parsing
python-docx==1.1.2
beautifulsoup4==4.12.3

# AI / LLM
anthropic==0.42.0
langchain==0.3.14
langgraph==0.2.59
langchain-anthropic==0.3.4

# Database (PostgreSQL + pgvector)
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pgvector==0.3.6
alembic==1.14.0
